// src/app/user-accounts/investors/details/profile/_comps/BackgroundForm/type.ts
export interface IBackgroundProps {
  data: {
    investmentsExperience: string;
    socialStatus: string;
    creditRisk: string;
    employmentStatus: string;
    employerName: string;
    employerAddress: string;
    countryOfEmployment: string;
    jobTitle: string;
    employmentIndustry: string;
    annualSalaryRange: string;
  };
  investorId: string;
}

export interface IFormData {
  investmentsExperience: string;
  socialStatus: string;
  creditRisk: string;
  employmentStatus: string;
  employerName: string;
  employerAddress: string;
  countryOfEmployment: string;
  jobTitle: string;
  employmentIndustry: string;
  annualSalaryRange: string;
}